Modified the README.md to include a second example
Modified the README.md to include a second example where the account SID and
auth token are embedded in the code rather than stored as environmental
variables. This would allow a developer to compile a self-contained executable
that could be run by end users without the need to create or understand
environment variables. Also modified both examples to define sending and
receiving phone number bindings to make it easier for the developer to
read/configure.