Fix Travis CI failures
I saw some broken log output that looked familiar to me, so this PR disables
`filter_secrets` (we do our best not to leak these, even without Travis's help,
so we'll just have to continue ensuring we don't `echo` any secrets). We also
have some timeouts for normal `cabal build`s, so I added the workaround from
[here](https://github.com/hvr/multi-ghc-travis#known-issues).

Also, in our test, we POST to /Messages using the same To and From number. I
guess this doesn't work anymore (I thought it used to), so I set the To number
to something else.